Added Higned SD, rearranged silk
Added new hinged MicroSD for stability purposes, this sd card will be operating in a high g, high vibration enviornment and a spring ejection slot is not adequate. I also rearranged the labels on the silk screen as the headers were blocking a few. The SD card is also seated closer to the center, hopefully so the card no loner sticks out past the board dimensions.